% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhondt.R
\name{dhondt}
\alias{dhondt}
\title{dHondt function (individual entries)}
\usage{
dhondt(n, a1, a2, a3 = 0, a4 = 0, a5 = 0, a6 = 0, a7 = 0, a8 = 0,
  a9 = 0, a10 = 0, a11 = 0, a12 = 0, vis = TRUE)
}
\arguments{
\item{n}{Number of seats to apportion.}

\item{vis}{Default TRUE. Do you want to display the output or not?}

\item{a}{Population in the district.}
}
\description{
This function automatically apportions seats proportional to the population.
}

