% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhondtV.R
\name{dhondtV}
\alias{dhondtV}
\title{dHondt function (Vector entry)}
\usage{
dhondtV(n, x, vis = TRUE)
}
\arguments{
\item{n}{Number of seats to apportion.}

\item{x}{Vector of population inputs.}

\item{vis}{Default TRUE. Do you want to display the output or not?}
}
\description{
This function automatically apportions seats proportional to the population.
}

